refactor: remove unused ClusterRole
The training-operator charm is always deployed using trust because it needs
elevated permissions to apply k8s resources using lightkube, so there is not
an actual need for a Role or a ClusterRole. This commit removes the unused
ClusterRole, keeping just the aggregated roles (for them to be applied to the
Kubeflow Profile namespace) when needed.