refactor: deploy the training-operator with kubernetes resources (#161)
* refactor: deploy the training-operator with kubernetes resources

This commit refactors the way the training-operator is deployed, as instead of using
a sidecar container that runs the workload, we are now applying the Deployment and all
the Kubernetes resources required by the training-operator controller to be able to mange
training resources.
We are introducing this change in preparation for the upcoming 1.8 version, as it introduces the
hard dependency on a Kubernetes Secret mounted in a volume for the training-operator workload
to start. For more details please refer to canonical/training-operator#159.